theoretically fix "queen died" popup, add locker emptying & unbuckling to prevent people getting deleted. still a bug with unbuckled people being inserted into the nest
