+++ Content.Server/_Impstation/Replicator/ReplicatorNestSystem.cs
@@ -6,21 +6,29 @@ using Content.Server._Impstation.Administration.Components;
 using Content.Server.Actions;
 using Content.Server.Announcements.Systems;
 using Content.Server.Audio;
+using Content.Server.Buckle.Systems;
 using Content.Server.GameTicking;
 using Content.Server.Pinpointer;
 using Content.Server.Popups;
+using Content.Server.Storage.Components;
+using Content.Server.Storage.EntitySystems;
 using Content.Server.Stunnable;
 using Content.Shared._Impstation.Replicator;
 using Content.Shared.Actions;
 using Content.Shared.Audio;
+using Content.Shared.Buckle.Components;
 using Content.Shared.Destructible;
 using Content.Shared.Inventory;
 using Content.Shared.Mind.Components;
+using Content.Shared.Mobs.Components;
 using Content.Shared.Mobs.Systems;
 using Content.Shared.Movement.Events;
+using Content.Shared.Movement.Pulling.Components;
+using Content.Shared.Movement.Pulling.Systems;
 using Content.Shared.Pinpointer;
 using Content.Shared.StepTrigger.Systems;
 using Content.Shared.Stunnable;
+using Content.Shared.Throwing;
 using Content.Shared.Whitelist;
 using Robust.Server.Containers;
 using Robust.Server.GameObjects;
@@ -37,6 +45,7 @@ public sealed class ReplicatorNestSystem : SharedReplicatorNestSystem
     [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
 
+    [Dependency] private readonly SharedReplicatorNestSystem _sharedNest = default!;
     [Dependency] private readonly ActionContainerSystem _actionContainer = default!;
     [Dependency] private readonly ActionsSystem _actions = default!;
     [Dependency] private readonly ContainerSystem _containerSystem = default!;
@@ -44,12 +53,16 @@ public sealed class ReplicatorNestSystem : SharedReplicatorNestSystem
     [Dependency] private readonly NavMapSystem _navMap = default!;
     [Dependency] private readonly PopupSystem _popup = default!;
     [Dependency] private readonly StunSystem _stun = default!;
-    [Dependency] private readonly TransformSystem _transform = default!;
+    [Dependency] private readonly TransformSystem _xform = default!;
     [Dependency] private readonly InventorySystem _inventory = default!;
     [Dependency] private readonly PinpointerSystem _pinpointer = default!;
     [Dependency] private readonly AmbientSoundSystem _ambientSound = default!;
     [Dependency] private readonly EntityWhitelistSystem _whitelist = default!;
     [Dependency] private readonly AnnouncerSystem _announcer = default!;
+    [Dependency] private readonly PullingSystem _pulling = default!;
+    [Dependency] private readonly ThrowingSystem _throwing = default!;
+    [Dependency] private readonly EntityStorageSystem _entStorage = default!;
+    [Dependency] private readonly BuckleSystem _buckle = default!;
 
     public override void Initialize()
     {
@@ -58,6 +71,7 @@ public sealed class ReplicatorNestSystem : SharedReplicatorNestSystem
         SubscribeLocalEvent<ReplicatorNestComponent, MapInitEvent>(OnMapInit);
         SubscribeLocalEvent<ReplicatorNestComponent, EntRemovedFromContainerMessage>(OnEntRemoved);
         SubscribeLocalEvent<ReplicatorNestComponent, StepTriggerAttemptEvent>(OnStepTriggerAttempt);
+        SubscribeLocalEvent<ReplicatorNestComponent, StepTriggeredOffEvent>(OnStepTriggered);
         SubscribeLocalEvent<ReplicatorNestFallingComponent, UpdateCanMoveEvent>(OnUpdateCanMove);
         SubscribeLocalEvent<ReplicatorNestComponent, DestructionEventArgs>(OnDestroyed);
         SubscribeLocalEvent<RoundEndTextAppendEvent>(OnRoundEndTextAppend);
@@ -124,6 +138,60 @@ public sealed class ReplicatorNestSystem : SharedReplicatorNestSystem
     private void OnStepTriggerAttempt(Entity<ReplicatorNestComponent> ent, ref StepTriggerAttemptEvent args)
     {
         args.Continue = true;
+    }
+
+    private void OnStepTriggered(Entity<ReplicatorNestComponent> ent, ref StepTriggeredOffEvent args)
+    {
+        // dont accept if they are already falling
+        if (HasComp<ReplicatorNestFallingComponent>(args.Tripper))
+            return;
+
+        // *reject* if blacklisted
+        if (_whitelist.IsBlacklistPass(ent.Comp.Blacklist, args.Tripper))
+        {
+            if (TryComp<PullableComponent>(args.Tripper, out var pullable) && pullable.BeingPulled)
+                _pulling.TryStopPull(args.Tripper, pullable);
+
+            var xform = Transform(ent);
+            var xformQuery = GetEntityQuery<TransformComponent>();
+            var worldPos = _xform.GetWorldPosition(xform, xformQuery);
+
+            var direction = _xform.GetWorldPosition(args.Tripper, xformQuery) - worldPos;
+            _throwing.TryThrow(args.Tripper, direction * 10, 7, ent, 0);
+            return;
+        }
+
+        var isReplicator = HasComp<ReplicatorComponent>(args.Tripper);
+
+        // Allow dead replicators regardless of current level.
+        if (TryComp<MobStateComponent>(args.Tripper, out var mobState) && isReplicator && _mobState.IsDead(args.Tripper))
+        {
+            _sharedNest.StartFalling(ent, args.Tripper);
+            return;
+        }
+
+        // Don't allow living beings. If you want those sweet bonus points, you have to kill.
+        if (mobState != null && _mobState.IsAlive(args.Tripper))
+            return;
+
+        // if the ent is a container, all its contents go in the hole
+        if (TryComp<EntityStorageComponent>(args.Tripper, out var entStorage))
+        {
+            _entStorage.EmptyContents(args.Tripper, entStorage);
+        }
+
+        if (TryComp<StrapComponent>(args.Tripper, out var strapComp) && strapComp.BuckledEntities.Count > 0)
+        {
+            foreach (var buckled in strapComp.BuckledEntities)
+            {
+                if (!TryComp<BuckleComponent>(buckled, out var buckleComp))
+                {
+                    _buckle.Unbuckle((args.Tripper, buckleComp), null);
+                }
+            }
+        }
+
+        _sharedNest.StartFalling(ent, args.Tripper);
     }
 
     private void OnUpdateCanMove(Entity<ReplicatorNestFallingComponent> ent, ref UpdateCanMoveEvent args)
@@ -262,7 +330,7 @@ public sealed class ReplicatorNestSystem : SharedReplicatorNestSystem
             i++;
             var pointsStorage = ent.Comp;
             var location = "Unknown";
-            var mapCoords = _transform.ToMapCoordinates(Transform(ent).Coordinates);
+            var mapCoords = _xform.ToMapCoordinates(Transform(ent).Coordinates);
             if (_navMap.TryGetNearestBeacon(mapCoords, out var beacon, out _) && beacon?.Comp.Text != null)
                 location = beacon?.Comp.Text!;
