+++ Content.Server/_Impstation/Replicator/ReplicatorSystem.cs
@@ -175,15 +175,17 @@ public sealed class ReplicatorSystem : EntitySystem
 
     private void OnMobStateChanged(Entity<ReplicatorComponent> ent, ref MobStateChangedEvent args)
     {
-        if (args.NewMobState != MobState.Critical || args.NewMobState != MobState.Dead)
+        if (args.NewMobState != MobState.Critical | args.NewMobState != MobState.Dead)
             return;
 
         _appearance.SetData(ent, ReplicatorVisuals.Combat, false);
 
-        if (ent.Comp.Queen)
+        if (HasComp<ReplicatorSignComponent>(ent))
         {
             foreach (var (uid, comp) in ent.Comp.RelatedReplicators)
+            {
                 _popup.PopupEntity(Loc.GetString(comp.QueenDiedMessage), uid, uid, PopupType.LargeCaution);
+            }
         }
     }
